% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{avoid_gaps_update}
\alias{avoid_gaps_update}
\title{Is is a function designed to remove values <= to 'gaps_threshold'. 
Nucleotides local and global positions, bins, size of regions/genes and exons
will be recalculated. To use on metagene's table during RNA-seq analysis. 
Not made for ChIP-Seq analysis or to apply on matagene's data_frame. A
similar function is implemented in produce_data_frame() with same arguments.}
\usage{
avoid_gaps_update(table, bam_name, gaps_threshold = 0)
}
\arguments{
\item{table}{A data.table from produce_table(...) function of metagene.}

\item{bam_name}{A reference bam_name to allow the same removal 
(position in bam) of values for other bam file.}

\item{gaps_threshold}{A threshold under which values will be removed.}
}
\value{
A data.table with values <= to 'gaps_threshold' removed
}
\description{
Is is a function designed to remove values <= to 'gaps_threshold'. 
Nucleotides local and global positions, bins, size of regions/genes and exons
will be recalculated. To use on metagene's table during RNA-seq analysis. 
Not made for ChIP-Seq analysis or to apply on matagene's data_frame. A
similar function is implemented in produce_data_frame() with same arguments.
}
\examples{
\dontrun{
 bam_files <- c(
 system.file("extdata/c_al4_945MLM_demo_sorted.bam", package="metagene"),
 system.file("extdata/c_al3_362PYX_demo_sorted.bam", package="metagene"),
 system.file("extdata/n_al4_310HII_demo_sorted.bam", package="metagene"),
 system.file("extdata/n_al3_588WMR_demo_sorted.bam", package="metagene"))
 region <- c(
 system.file("extdata/ENCFF355RXX_DPM1less.bed", package="metagene"),
 system.file("extdata/ENCFF355RXX_NDUFAB1less.bed", package="metagene"),
 system.file("extdata/ENCFF355RXX_SLC25A5less.bed", package="metagene"))
 mg <- metagene$new(regions = region, bam_files = bam_files, 
                                                           assay = 'rnaseq')
 tab <- mg$get_table()
 tab <- avoid_gaps_update(tab, 
        bam_name = 'c_al4_945MLM_demo_sorted', gaps_threshold = 10)
}

}
