% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_handler.R
\docType{data}
\name{Bam_Handler}
\alias{Bam_Handler}
\title{A class to manage BAM files.}
\format{A BAM manager}
\usage{
Bam_Handler
}
\value{
\code{Bam_Handler$new} returns a \code{Bam_Handler} object which contains
coverage related information for every BAM files.
}
\description{
This class will allow to load, convert and normalize alignments and regions
files/data.
}
\section{Constructor}{

\describe{
   \item{}{\code{bh <- Bam_Handler$new(bam_files, cores = SerialParam())}}
   \item{bam_files}{A \code{vector} of BAM filenames. The BAM files must be
                   indexed. i.e.: if a file is named file.bam, there must
                   be a file named file.bam.bai or file.bai in the same 
                   directory.}
   \item{cores}{The number of cores available to parallelize the analysis.
               Either a positive integer or a \code{BiocParallelParam}.
               Default: \code{SerialParam()}.}
}

\code{Bam_Handler$new} returns a \code{Bam_Handler} object that contains
and manages BAM files. Coverage related information as alignment count can
be obtain by using this object.
}

\section{Methods}{

\describe{
   \item{}{\code{bh$get_aligned_count(bam_file)}}
   \item{bam_file}{The name of the BAM file.}
}
\describe{
   \item{}{\code{bg$get_bam_name(bam_file)}}
   \item{bam_file}{The name of the BAM file.}
}
\describe{
   \item{}{\code{bh$get_rpm_coefficient(bam_file)}}
   \item{bam_file}{The name of the BAM file.}
}
\describe{
   \item{}{\code{bh$index_bam_files(bam_files)}}
   \item{bam_files}{A \code{vector} of BAM filenames.}
}
\describe{
   \item{}{\code{bh$get_bam_files()}}
}
\describe{
   \item{}{\code{bh$get_coverage(bam_file, regions)
                               force_seqlevels = FALSE)}}
   \item{bam_file}{The name of the BAM file.}
   \item{regions}{A not empty \code{GRanges} object.}
   \item{force_seqlevels}{If \code{TRUE}, Remove regions that are not found
                   in bam file header. Default: \code{FALSE}. TRUE and FALSE
                   respectively correspond to pruning.mode = "coarse" 
                   and "error" in ?seqinfo.}
}
\describe{
   \item{}{\code{bh$get_normalized_coverage(bam_file, regions)
                               force_seqlevels = FALSE)}}
   \item{bam_file}{The name of the BAM file.}
   \item{regions}{A not empty \code{GRanges} object.}
   \item{force_seqlevels}{If \code{TRUE}, Remove regions that are not found
                   in bam file header. Default: \code{FALSE}. TRUE and FALSE
                   respectively correspond to pruning.mode = "coarse" 
                   and "error" in ?seqinfo.}
}
\describe{
   \item{}{\code{bh$get_noise_ratio(chip_bam_file, input_bam_file)}}
   \item{chip_bam_file}{The path to the chip bam file.}
   \item{input_bam_file}{The path to the input (control) bam file.}
}
}

\examples{
bam_file <- get_demo_bam_files()[1]
bh <- metagene:::Bam_Handler$new(bam_files = bam_file)
bh$get_aligned_count(bam_file)

}
\keyword{datasets}
