% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/demo.R
\name{get_demo_bam_files}
\alias{get_demo_bam_files}
\title{Get BAM filenames for demo}
\usage{
get_demo_bam_files()
}
\value{
A vector of BAM filenames
}
\description{
Get BAM filenames for demo
}
\examples{
bam_files <- get_demo_bam_files()
}

