% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{get_demo_design}
\alias{get_demo_design}
\title{Get a demo design object}
\usage{
get_demo_design()
}
\value{
A \code{data.frame} corresponding to a valid design.
}
\description{
Get a demo design object
}
\examples{
mg <- get_demo_design()
}
