% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{get_demo_metagene}
\alias{get_demo_metagene}
\title{Get a demo metagene object}
\usage{
get_demo_metagene()
}
\value{
A metagene object
}
\description{
Get a demo metagene object
}
\examples{
mg <- get_demo_metagene()
}

