% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_demo_regions}
\alias{get_demo_regions}
\title{Get regions filenames for demo}
\usage{
get_demo_regions()
}
\value{
A vector of regions filenames
}
\description{
Get regions filenames for demo
}
\examples{
regions <- get_demo_regions()
}

