% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{get_promoters_txdb}
\alias{get_promoters_txdb}
\title{Extract Entrez genes promoters from TxDb object.}
\usage{
get_promoters_txdb(txdb, upstream = 1000, downstream = 1000)
}
\arguments{
\item{txdb}{A valid \code{TxDb} object.}

\item{upstream}{The number of nucleotides upstream of TSS.}

\item{downstream}{The number of nucleotides downstream of TSS.}
}
\value{
A \code{GRanges} object that contains the promoters infos.
}
\description{
Extract Entrez genes promoters from TxDb object.
}
\examples{
\dontrun{
    # require(TxDb.Hsapiens.UCSC.hg19.knownGene)
    txdb <- TxDb.Hsapiens.UCSC.hg19.knownGene
    promoters_hg19 <- get_promoters_txdb(txdb)
}

}
