% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{permutation_test}
\alias{permutation_test}
\title{Perform a permutation test on 2 matrices}
\usage{
permutation_test(matrix1, matrix2, sample_size, sample_count, FUN, ...)
}
\arguments{
\item{matrix1}{The first matrix.}

\item{matrix2}{The second matrix.}

\item{sample_size}{The number of element to draw for each matrix.}

\item{sample_count}{The number of permutations.}

\item{FUN}{The function to use to compare the 2 matrices. First two params
must be \code{numeric} vector and the return must be a single
\code{numeric} value.}

\item{...}{Extra param for \code{FUN}.}
}
\value{
A \code{vector} of numeric corresponding to the result of each
        permutation.
}
\description{
The goal of this function is to calculate the values of a test performed
by \code{FUN} after each of \code{sample_count} permutations.
}
\details{
Each round of the permutation test, two new matrices will be randomly
sampled from using the combination of the two original matrices. The means
of each columns will be calculated to produce the vectors that will be sent
\code{FUN}.
}
\examples{
\dontrun{
# Get some matrices
mg <- get_demo_metagene()
mg$produce_matrices()
m1 <- mg$get_matrices()$list1$align1_rep1$input
m2 <- mg$get_matrices()$list1$align2_rep1$input

# Perform permutation test
sample_size <- min(nrow(m1), nrow(m2))
FUN = function(a, b) { mean(a) - mean(b) } # Dummy function for demo purpose
# A sample_count >= 1000 should be used in a real analysis
permutation_results <- permutation_test(m1, m2, sample_size = sample_size,
                                        sample_count = 10, FUN = FUN)
}

}
