% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_metagene.R
\name{plot_metagene}
\alias{plot_metagene}
\title{Produce a metagene plot}
\usage{
plot_metagene(df)
}
\arguments{
\item{df}{a \code{data.frame} obtained with the \code{get_data_frame}
function. Must have the following columns: "region", "design", "bin", 
"value", "qinf" and "qsup".}
}
\value{
A `ggplot` object.
}
\description{
Produce a metagene plot
}
\examples{
region <- get_demo_regions()[1]
bam_file <- get_demo_bam_files()[1]
mg <- metagene$new(regions = region, bam_files = bam_file)
mg$produce_data_frame()
df <- mg$get_data_frame()
p <- plot_metagene(df)
}
