% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{promoters_hg18}
\alias{promoters_hg18}
\title{Promoters regions of hg18 Entrez genes.}
\format{A \code{GRanges} object with 19742 ranges.}
\usage{
promoters_hg18
}
\value{
A \code{GRanges}.
}
\description{
Each regions have a width of 2000 nucleotide centered at the transcription
start site.
}
\examples{
data(promoters_hg18)
}
\seealso{
\code{\link{get_promoters_txdb}}
}
\keyword{datasets}

