% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{promoters_hg19}
\alias{promoters_hg19}
\title{Promoters regions of hg19 Entrez genes.}
\format{A \code{GRanges} object with 23056 ranges.}
\usage{
promoters_hg19
}
\value{
A \code{GRanges}.
}
\description{
Each regions have a width of 2000 nucleotide centered at the transcription
start site.
}
\examples{
data(promoters_hg19)
}
\seealso{
\code{\link{get_promoters_txdb}}
}
\keyword{datasets}
