% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{promoters_mm10}
\alias{promoters_mm10}
\title{Promoters regions of mm10 Entrez genes.}
\format{A \code{GRanges} object with 23653 ranges.}
\usage{
promoters_mm10
}
\value{
A \code{GRanges}.
}
\description{
Each regions have a width of 2000 nucleotide centered at the transcription
start site.
}
\examples{
data(promoters_mm10)
}
\seealso{
\code{\link{get_promoters_txdb}}
}
\keyword{datasets}

