% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{promoters_mm9}
\alias{promoters_mm9}
\title{Promoters regions of mm9 Entrez genes.}
\format{A \code{GRanges} object with 21677 ranges.}
\usage{
promoters_mm9
}
\value{
A \code{GRanges}.
}
\description{
Each regions have a width of 2000 nucleotide centered at the transcription
start site.
}
\examples{
data(promoters_mm9)
}
\seealso{
\code{\link{get_promoters_txdb}}
}
\keyword{datasets}
