\name{scaleVector}
\alias{scaleVector}

\title{
  Scale the values of a vector to fit with predetermined size
}

\usage{
  scaleVector(
    values,
    domain)
}

\arguments{
  \item{values}{the values to scale}

  \item{domain}{the range to fit the value to}
}

\value{
  \code{scaleVector} returns a \code{vector} with the scaled data
}

\author{
  Charles Joly Beauparlant
  <Charles.Joly-Beauparlant@crchul.ulaval.ca>
}
