% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{write_bed_file_filter_result}
\alias{write_bed_file_filter_result}
\title{Transforms the bed_file_filter function output into a file.BED readable by
                                                               metagene.}
\usage{
write_bed_file_filter_result(bed_file_filter_result, file = "file_name",
  path = "./")
}
\arguments{
\item{bed_file_filter_result}{A \code{GRanges} object : the output of 
bed_file_filter function.}

\item{file}{the name of the output file without the extension}

\item{path}{The path where the function will write the file}
}
\value{
output of write function
}
\description{
Transforms the bed_file_filter function output into a file.BED readable by
                                                               metagene.
}
\examples{
\dontrun{
require(EnsDb.Hsapiens.v86)
edb <- EnsDb.Hsapiens.v86
quantification_files <- 'file_path'
ebgwot <- exon_by_gene_with_observed_transcripts(edb, 
                                                   quantification_files)
bed_file_content_gr <- GRanges("chr16",ranges = IRanges(start=23581002, 
                                                           end=23596356))
bffr <- bed_file_filter(ebgwot, bed_file_content_gr)
write_bed_file_filter_result(bffr, file='test','./')
}

}
