% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagene.R
\docType{data}
\name{metagene}
\alias{metagene}
\title{A class to manage metagene analysis.}
\format{A metagene experiment manager}
\usage{
metagene
}
\value{
\code{metagene$new} returns a \code{metagene} object which contains the
normalized coverage values for every regions and for every BAM files.
}
\description{
This class will allow to load, convert and normalize alignments and regions
files/data. Once the data is ready, the user can then chose to produce
metagene plots on the data (or a subset of the data).
}
\section{Constructor}{

\describe{
   \item{}{\code{mg <- metagene$new(regions, bam_files, padding_size = 0,
                           cores = SerialParam(), verbose = FALSE,
                           force_seqlevels = FALSE)}}
   \item{regions}{Either a \code{vector} of BED, narrowPeak or broadPeak
                   filenames, a \code{GRanges} object or a 
                   \code{GRangesList} object.}
   \item{bam_files}{A \code{vector} of BAM filenames. The BAM files must be
                   indexed. i.e.: if a file is named file.bam, there must
                   be a file named file.bam.bai or file.bai in the same 
                   directory.}
   \item{padding_size}{The regions will be extended on each side by the
                       value of this parameter. The padding_size must be a
                       non-negative integer. Default = 0.}
   \item{cores}{The number of cores available to parallelize the analysis.
               Either a positive integer or a \code{BiocParallelParam}.
               Default: \code{SerialParam()}.}
   \item{verbose}{Print progression of the analysis. A logical constant.
                   Default: \code{FALSE}.}
   \item{force_seqlevels}{If \code{TRUE}, Remove regions that are not found
               in bam file header. Default: \code{FALSE}. TRUE and FALSE
               respectively correspond to pruning.mode = "coarse"
               and "error" in ?seqinfo.}
     \item{paired_end}{If \code{TRUE}, metagene will deal with paired-ended 
                 data. If \code{FALSE}, single-ended data are expected. 
                 Default: \code{FALSE}}
     \item{assay}{\code{'chipseq'} or \code{'rnaseq'}, the two available 
                 options. Default: \code{'chipseq'}}
}

   \code{metagene$new} returns a \code{metagene} object that contains the
       coverages for every BAM files in the regions from the \code{regions}
       param.
}

\examples{
region <- get_demo_regions()[1]
bam_file <- get_demo_bam_files()[1]
mg <- metagene$new(regions = region, bam_files = bam_file)
\dontrun{
   df <- metagene$plot()
}

}
\keyword{datasets}
