% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{metagene}
\alias{metagene}
\title{A class to manage metagene analysis.}
\format{A metagene experiment manager}
\usage{
metagene
}
\value{
\code{metagene$new} returns a \code{metagene} object which contains the
normalized coverage values for every regions and for every BAM files.
}
\description{
This class will allow to load, convert and normalize alignments and regions
files/data. Once the data is ready, the user can then chose to produce
metagene plots on the data (or a subset of the data).
}
\section{Constructor}{

\describe{
  \item{}{\code{mg <- metagene$new(regions, bam_files, padding_size = 0,
                             cores = SerialParam(), verbose = FALSE)}}
  \item{regions}{Either a \code{list} of bed filenames, a \code{GRanges}
                 object or a \code{GRangesList} object.}
  \item{bam_files}{A list of bam filenames. The bam files must be indexed.
                   i.e.: if a file is named file.bam, there must be a file
                   named file.bam.bai in the same directory.}
  \item{padding_size}{The regions will be extended on each side by the value
                      of this parameter. The padding_size must be a
                      non-negative integer. Default = 0.}
  \item{cores}{The number of cores available to parallelize the analysis.
               Either a positive integer or a \code{BiocParallelParam}.
               Default: SerialParam().}
  \item{verbose}{Print progression of the analysis. A logical constant.
                 Default: \code{FALSE}.}
}

 \code{metagene$new} returns a \code{metagene} object that contains the
 coverages for every bam files in the regions from the \code{regions} param.
}

\section{Methods}{

\describe{
  \item{}{\code{df <- mg$plot(design = NULL, regions_group = NULL,
              bin_size = 100, alpha = 0.05, sample_count = 1000)}}
  \item{design}{A \code{data.frame} that describe to experiment to plot. The
                first column must be the filenames. The other columns (at
                least one is required) represent how the files should be
                grouped. All the files in the same group will be combined
                before doing the statistical analysis. For each column,
                there will be one line with its ribon in the metagene plot.
                Default = NULL.

                0: Do not use file.
                1: File is input.
                2: File is control.}
  \item{region_group}{A list of region names to include in the analysis. If
                     \code{NULL}, all the regions used when creating the
                     \code{metagene} object will be used.
                     Default: \code{NULL}}
  \item{bin_size}{The size of bin to use before calculating the statistics.
                  larger \code{bin_size} will reduce the calculation time
                  and produce smoother curves. Default = 100.}
  \item{alpha}{The condifence interval (CI) to represent with a ribbon. Must
               be a value between 0 and 1. With a value of 0.05, the ribbon
               will represent 95% of the estimated values of the means.
               Default = 0.05.}
  \item{sample_count}{The number of draw to do during bootstrap analysis.
                      Default: 1000.}
}
\describe{
  \item{}{\code{mg$export(bam_file, region, file}}
  \item{bam_file}{The name of the bam file to export.}
  \item{region}{The name of the region to export.}
  \item{file}{The name of the ouput file.}
}
\describe{
  \item{}{\code{mg$heatmap(region, bam_file, bin_size}}
  \item{region}{The name of the region to export.}
  \item{bam_file}{The name of the bam file to export.}
  \item{bin_size}{The size of the bin to produce before creating heatmap.}
}
}
\examples{
regions <- get_demo_regions()
 bam_files <- get_demo_bam_files()
 mg <- metagene$new(regions, bam_files)
 \dontrun{
 df <- metagene$plot()}
}
\keyword{datasets}

