% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metagene.R
\docType{data}
\name{metagene}
\alias{metagene}
\title{A class to manage metagene analysis.}
\format{A metagene experiment manager}
\usage{
metagene
}
\value{
\code{metagene$new} returns a \code{metagene} object which contains the
normalized coverage values for every regions and for every BAM files.
}
\description{
This class will allow to load, convert and normalize alignments and regions
files/data. Once the data is ready, the user can then chose to produce
metagene plots on the data (or a subset of the data).
}
\section{Constructor}{

\describe{
  \item{}{\code{mg <- metagene$new(regions, bam_files, padding_size = 0,
                             cores = SerialParam(), verbose = FALSE,
                             force_seqlevels = FALSE)}}
  \item{regions}{Either a \code{vector} of BED, narrowPeak or broadPeak
                 filenames, a \code{GRanges} object or a \code{GRangesList}
                 object.}
  \item{bam_files}{A \code{vector} of BAM filenames. The BAM files must be
                   indexed. i.e.: if a file is named file.bam, there must be
                   a file named file.bam.bai in the same directory.}
  \item{padding_size}{The regions will be extended on each side by the value
                      of this parameter. The padding_size must be a
                      non-negative integer. Default = 0.}
  \item{cores}{The number of cores available to parallelize the analysis.
               Either a positive integer or a \code{BiocParallelParam}.
               Default: \code{SerialParam()}.}
  \item{verbose}{Print progression of the analysis. A logical constant.
                 Default: \code{FALSE}.}
  \item{force_seqlevels}{If \code{TRUE}, Remove regions that are not found in
                         bam file header. Default: \code{FALSE}.}
}

 \code{metagene$new} returns a \code{metagene} object that contains the
 coverages for every BAM files in the regions from the \code{regions} param.
}

\section{Methods}{

\describe{
  \item{}{\code{df <- mg$plot(design = NULL, regions_group = NULL,
              bin_size = 100, alpha = 0.05, sample_count = 1000,
              title = NULL, flip_regions)}}
  \item{design}{A \code{data.frame} that describe to experiment to plot. The
                first column must be the existing BAM filenames or the BAM
                name that was used when creating the metagene object. The
                other columns (at least one is required) represent how the
                files should be grouped. All those colums should be in
                numeric format. All the files in the same group will be
                combined before doing the statistical analysis. For each
                column, there will be one line with its ribon in
                the metagene plot. At least one file should be selected (not
                zero). If \code{NULL}, all BAM files passed as argument
                during \code{metagene} object creation will be used to
                create a default design.
                Default = \code{NULL}.

                0: Do not use file.
                1: File is input.
                2: File is control.}
  \item{regions_group}{A \code{list} or a \code{vector} of region names to
                     include in the analysis. If \code{NULL}, all the
                     regions used when creating the
                     \code{metagene} object will be used.
                     Default: \code{NULL}}
  \item{bin_size}{The size of bin to use before calculating the statistics.
                  larger \code{bin_size} will reduce the calculation time
                  and produce smoother curves. It must be a positive
                  integer. Default = 100.}
  \item{alpha}{The condifence interval (CI) to represent with a ribbon. Must
               be a value between 0 and 1. With a value of 0.05, the ribbon
               will represent 95% of the estimated values of the means.
               Default = 0.05.}
  \item{sample_count}{The number of draw to do during bootstrap analysis.
                      Default: 1000.}
  \item{title}{A title to add to the graph. If \code{NULL}, will be
                      automatically created. Default: NULL}
  \item{flip_regions}{Should regions on negative strand be flip_regions?
                      Default: \code{FALSE}.}
}
\describe{
  \item{}{\code{mg$produce_matrices(select_regions, design, bin_count,
                bin_size, noise_removal, normalization, flip_regions}}
  \item{select_regions}{A character vector representing corresponding to the
                        regions to include in the metagene plot. The names
                        of the regions must correspond to the names used for
                        the initialization (\code{new}) of the metagene
                        object. \code{NA} can be used to keep previous
                        select_regions values. \code{NULL} can be used to
                        keep all regions. Default: \code{NA}.}
  \item{design}{A \code{data.frame} that describe to experiment to plot. See
                \code{new} function for more details. \code{NA} can be used
                keep previous design value. Default: \code{NA}.}
  \item{bin_count}{The number of bin to create. \code{NA} can be used to
                   keep previous bin_count value. If both bin_count and
                   bin_size are \code{NA}, a bin_count value of 100 will be
                   used. Default: \code{NA}.}
  \item{bin_size}{The size of bin to create. Can only be used if all the
                  regions have the same size. \code{NA} can be used to keep
                  previous bin_size value. If both bin_count and bin_size
                  are \code{NA}, a bin_count value of 100 will be used.
                  Default: \code{NA}.}
  \item{noise_removal}{The algorithm to use to remove control(s). Possible
                       default, value is \code{NULL}. Use \code{NA} keep
                       previous \code{noise_removal} value (i.e. if
                       \code{produce_matrices} was called before). See
                       Liand and Keles 2012 for the NCIS algorithm.}
  \item{normalization}{The algorithm to use to normalize samples. Possible
                       default, value is \code{NULL}. Use \code{NA} keep
                       previous \code{normalization} value (i.e. if
                       \code{produce_matrices} was called before).}
  \item{flip_regions}{Should regions on negative strand be flip_regions?
                      Default: \code{FALSE}.}
}
\describe{
  \item{}{\code{mg$export(bam_file, region, file)}}
  \item{bam_file}{The name of the bam file to export.}
  \item{region}{The name of the region to export.}
  \item{file}{The name of the ouput file.}
}
\describe{
  \item{}{\code{mg$heatmap(region, bam_file, bin_size)}}
  \item{region}{The name of the region to export.}
  \item{bam_file}{The name of the bam file to export.}
  \item{bin_size}{The size of the bin to produce before creating heatmap. It
                     must be a positive integer. Default: 10.}
}
\describe{
  \item{}{\code{mg$add_design(design = NULL)}}
  \item{design}{A \code{data.frame} that describe to experiment to plot. See
                \code{new} function for more details. \code{NA} can be used
                keep previous design value. Default: \code{NA}.}
}
\describe{
  \item{}{\code{mg$unflip_regions()}}
}

\describe{
  \item{}{\code{mg$flip_regions()}}
}
\describe{
  \item{}{\code{mg$unflip_regions()}}
}
}
\examples{
regions <- get_demo_regions()
 bam_files <- get_demo_bam_files()
 mg <- metagene$new(regions, bam_files)
 \dontrun{
 df <- metagene$plot()}
}
\keyword{datasets}

