\name{plot.graphic}
\alias{plot.graphic}

\title{
  Produce a plot with based on a data.frame
}

\usage{
  plot.graphic(
    DF,
    title)
}

\arguments{
  \item{DF}{The data frame produced by the plot.getDataFrame
    function}

  \item{title}{The title of the graph}
}

\value{
  \code{plot.graphic} plots a graph on the current device.
}

\author{
  Charles Joly Beauparlant
  <Charles.Joly-Beauparlant@crchul.ulaval.ca>
}
